fixed search after adding new entry
new entry was not added to recent search results and instead one of the entires is duplicated.

The cause was an incorrect use of map funciton.  Returning array in map function delegate resulted in array of arrays, which is not right.

returning a single entry with proper key fixed the issue